inventory/requirements/base.txt
Django==4.1.1
djangorestframework==3.14.0
psycopg2==2.9.3
nameko==2.14.1
drf_yasg==1.21.4